fixed bug not show correct info on navBar when loggin

+++ frontend/src/hooks/useExpenses.js
@@ -2,6 +2,7 @@
 
 import { useEffect, useState } from "react"
 import { apiService } from "../services/api.js"
+import { authService } from "../services/authService.js"
 
 export const useExpenses = () => {
     const type = 'expense'
@@ -18,6 +19,11 @@ export const useExpenses = () => {
     const loadExpenses = async () => {
         try{
             setLoading(true)
+                if (!authService.isAuthenticated()) {
+                    setError('User not authenticated. Please log in.');
+                    setExpenses([]);
+                    return;
+            }
             const data = await apiService.getAll(type)
             setExpenses(data)
             setError('')
@@ -30,6 +36,10 @@ export const useExpenses = () => {
 
     const createExpense = async (expense) => {
         try{
+            if (!authService.isAuthenticated()) {
+                setError('User not authenticated. Please log in.');
+                return;
+            }
             const  newExpense = await apiService.create(expense,type)
             setExpenses([...expenses, newExpense])
             setError('')
@@ -40,6 +50,10 @@ export const useExpenses = () => {
 
     const updateExpense = async (id, expense) => {
         try{
+            if (!authService.isAuthenticated()) {
+                setError('User not authenticated. Please log in.');
+                return;
+            }
             const updatedExpense = await apiService.update(id, expense,type)
             setExpenses(expenses.map(exp => exp.id === id ? updatedExpense : exp))
             setError('')
@@ -50,6 +64,10 @@ export const useExpenses = () => {
 
     const deleteExpense = async (id) => {
         try{
+            if (!authService.isAuthenticated()) {
+                setError('User not authenticated. Please log in.');
+                return;
+            }
             await apiService.delete(id,type)
             setExpenses(expenses.filter(item => item.id !== id))
             setError('')
@@ -60,6 +78,10 @@ export const useExpenses = () => {
 
     const filterExpenses = async (filters) => {
         try{
+            if (!authService.isAuthenticated()) {
+                setError('User not authenticated. Please log in.');
+                return;
+            }
             setLoading(true)
             const filtered = await apiService.filterExpenses(filters,type)
             setFilteredExpenses(filtered)
